worker: Move remote_setBuilderList() to deriving classes
